+++ framework/android/kodein-di-framework-android-core/src/main/java/org/kodein/di/android/closest.kt
@@ -49,8 +49,10 @@ private class LazyContextKodeinPropertyDelegateProvider(private val getContext: 
  * To be used on Android's `Context` classes, such as `Activity` or `Service`.
  */
 fun kodein(): KodeinPropertyDelegateProvider<Context> = ContextKodeinPropertyDelegateProvider()
-// Deprecated since 6.1
-@Deprecated("closestKodein has been renamed kodein", ReplaceWith("kodein()", "org.kodein.di.android.kodein"), DeprecationLevel.ERROR)
+
+/**
+ * Alias to `kodein`
+ */
 fun closestKodein() = kodein()
 
 /**
@@ -59,8 +61,10 @@ fun closestKodein() = kodein()
  * @param context The Android context to use to walk up the context hierarchy.
  */
 fun kodein(context: Context): KodeinPropertyDelegateProvider<Any?> = LazyContextKodeinPropertyDelegateProvider { context }
-// Deprecated since 6.1
-@Deprecated("closestKodein has been renamed kodein", ReplaceWith("kodein(context)", "org.kodein.di.android.kodein"), DeprecationLevel.ERROR)
+
+/**
+ * Alias to `kodein`
+ */
 fun closestKodein(context: Context) = kodein(context)
 
 /**
@@ -69,46 +73,58 @@ fun closestKodein(context: Context) = kodein(context)
  * @param getContext A function that returns the Android context to use to walk up the context hierarchy.
  */
 fun kodein(getContext: () -> Context): KodeinPropertyDelegateProvider<Any?> = LazyContextKodeinPropertyDelegateProvider(getContext)
-// Deprecated since 6.1
-@Deprecated("closestKodein has been renamed kodein", ReplaceWith("kodein(getContext)", "org.kodein.di.android.kodein"), DeprecationLevel.ERROR)
+
+/**
+ * Alias to `kodein`
+ */
 fun closestKodein(getContext: () -> Context) = kodein(getContext)
 
 /**
  * Returns the closest Kodein (or the app Kodein, if no closest Kodein could be found).
  */
 fun Fragment.kodein() = kodein { activity }
-// Deprecated since 6.1
-@Deprecated("closestKodein has been renamed kodein", ReplaceWith("kodein()", "org.kodein.di.android.kodein"), DeprecationLevel.ERROR)
+
+/**
+ * Alias to `kodein`
+ */
 fun Fragment.closestKodein() = kodein()
 
 /**
  * Returns the closest Kodein (or the app Kodein, if no closest Kodein could be found).
  */
 fun Dialog.kodein() = kodein { context }
-// Deprecated since 6.1
-@Deprecated("closestKodein has been renamed kodein", ReplaceWith("kodein()", "org.kodein.di.android.kodein"), DeprecationLevel.ERROR)
+
+/**
+ * Alias to `kodein`
+ */
 fun Dialog.closestKodein() = kodein()
 
 /**
  * Returns the closest Kodein (or the app Kodein, if no closest Kodein could be found).
  */
 fun View.kodein() = kodein { context }
-// Deprecated since 6.1
-@Deprecated("closestKodein has been renamed kodein", ReplaceWith("kodein()", "org.kodein.di.android.kodein"), DeprecationLevel.ERROR)
+
+/**
+ * Alias to `kodein`
+ */
 fun View.closestKodein() = kodein()
 
 /**
  * Returns the closest Kodein (or the app Kodein, if no closest Kodein could be found).
  */
 fun AbstractThreadedSyncAdapter.kodein() = kodein { context }
-// Deprecated since 6.1
-@Deprecated("closestKodein has been renamed kodein", ReplaceWith("kodein()", "org.kodein.di.android.kodein"), DeprecationLevel.ERROR)
+
+/**
+ * Alias to `kodein`
+ */
 fun AbstractThreadedSyncAdapter.closestKodein() = kodein()
 
 /**
  * Returns the closest Kodein (or the app Kodein, if no closest Kodein could be found).
  */
 fun Loader<*>.kodein() = kodein { context }
-// Deprecated since 6.1
-@Deprecated("closestKodein has been renamed kodein", ReplaceWith("kodein()", "org.kodein.di.android.kodein"), DeprecationLevel.ERROR)
+
+/**
+ * Alias to `kodein`
+ */
 fun Loader<*>.closestKodein() = kodein()