+++ framework/android/kodein-di-framework-android-x/src/main/java/org/kodein/di/android/x/closest.kt
@@ -11,17 +11,11 @@ import org.kodein.di.android.kodein
  * Returns the closest Kodein (or the app Kodein, if no closest Kodein could be found).
  */
 fun Fragment.kodein() = kodein { requireActivity() }
-// Deprecated since 6.1
-@Deprecated("closestKodein has been renamed kodein", ReplaceWith("kodein()", "org.kodein.di.android.x.kodein"), DeprecationLevel.ERROR)
-fun Fragment.closestKodein() = kodein()
 
 /**
  * Returns the closest Kodein (or the app Kodein, if no closest Kodein could be found).
  */
 fun Loader<*>.kodein() = kodein { context }
-// Deprecated since 6.1
-@Deprecated("closestKodein has been renamed kodein", ReplaceWith("kodein()", "org.kodein.di.android.x.kodein"), DeprecationLevel.ERROR)
-fun Loader<*>.closestKodein() = kodein()
 
 /**
  * Returns the closest Kodein (or the app Kodein, if no closest Kodein could be found).