+++ framework/tornadofx/kodein-di-framework-tornadofx-jvm/src/main/kotlin/org/kodein/di/tornadofx/scopes.kt
@@ -1,0 +1,16 @@
+package org.kodein.di.tornadofx
+
+import javafx.scene.*
+import org.kodein.di.bindings.*
+import tornadofx.*
+
+/**
+ * Scope that will leverage scoping instances on [Component]
+ * like [View] / [Fragment] / [ViewModel] / etc.
+ */
+object ComponentScope : WeakContextScope<Component>()
+/**
+ * Scope that will leverage scoping instances on [Node]
+ * Simply, every UI component
+ */
+object NodeScope : WeakContextScope<Node>()